+++ lib/credit_card_widget.dart
@@ -1,9 +1,10 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_credit_card/model/card_field.dart';
-import 'package:flutter_credit_card/model/card_patterns.dart';
+import 'package:flutter_credit_card/model/card_field_controller.dart';
+
 import 'package:flutter_credit_card/style/styles.dart';
+import 'package:flutter_credit_card/widget/animation_card.dart';
+import 'package:flutter_credit_card/widget/back_card.dart';
+import 'package:flutter_credit_card/widget/front_card.dart';
 import 'package:flutter_masked_text/flutter_masked_text.dart';
 
 class CreditCardWidget extends StatefulWidget {
@@ -55,6 +56,12 @@ class _CreditCardWidgetState extends State<CreditCardWidget>
   bool isAmex = false;
 
   @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
   void initState() {
     super.initState();
 
@@ -69,62 +76,19 @@ class _CreditCardWidgetState extends State<CreditCardWidget>
       _cvvController,
     ];
     _controllers.forEach((f) => f.addListener(() => setState(() {})));
+    backgroundGradientColor = cardGradient();
 
-    ///initialize the animation controller
-    controller = AnimationController(
+    controller = animationController(
       duration: widget.animationDuration,
       vsync: this,
     );
 
-    backgroundGradientColor = cardGradient();
-
-    ///Initialize the Front to back rotation tween sequence.
-    _frontRotation = TweenSequence<double>(
-      <TweenSequenceItem<double>>[
-        TweenSequenceItem<double>(
-          tween: Tween<double>(begin: 0.0, end: pi / 2)
-              .chain(CurveTween(curve: Curves.easeIn)),
-          weight: 50.0,
-        ),
-        TweenSequenceItem<double>(
-          tween: ConstantTween<double>(pi / 2),
-          weight: 50.0,
-        ),
-      ],
-    ).animate(controller);
-
-    _backRotation = TweenSequence<double>(
-      <TweenSequenceItem<double>>[
-        TweenSequenceItem<double>(
-          tween: ConstantTween<double>(pi / 2),
-          weight: 50.0,
-        ),
-        TweenSequenceItem<double>(
-          tween: Tween<double>(begin: -pi / 2, end: 0.0)
-              .chain(CurveTween(curve: Curves.easeOut)),
-          weight: 50.0,
-        ),
-      ],
-    ).animate(controller);
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
+    _frontRotation = frontRotation(controller);
+    _backRotation = backRotation(controller);
   }
 
   @override
   Widget build(BuildContext context) {
-    final double height = MediaQuery.of(context).size.height;
-    final double width = MediaQuery.of(context).size.width;
-    final Orientation orientation = MediaQuery.of(context).orientation;
-
-    ///
-    /// If uer adds CVV then toggle the card from front to back..
-    /// controller forward starts animation and shows back layout.
-    /// controller reverse starts animation and shows front layout.
-    ///
     if (widget.showBackView) {
       controller.forward();
     } else {
@@ -135,323 +99,19 @@ class _CreditCardWidgetState extends State<CreditCardWidget>
       children: <Widget>[
         AnimationCard(
           animation: _frontRotation,
-          child: buildFrontContainer(width, height, context, orientation),
+          child: FrontCard(
+            nameController: _nameController,
+            expiryController: _expiryController,
+            numberController: _numberController,
+          ),
         ),
         AnimationCard(
           animation: _backRotation,
-          child: buildBackContainer(width, height, context, orientation),
+          child: BackCard(
+              cvvController: _cvvController,
+              numberController: _numberController),
         ),
       ],
-    );
-  }
-
-  ///
-  /// Builds a back container containing cvv
-  ///
-  Container buildBackContainer(
-    double width,
-    double height,
-    BuildContext context,
-    Orientation orientation,
-  ) {
-    final TextStyle defaultTextStyle = Theme.of(context).textTheme.title.merge(
-          TextStyle(
-            color: Colors.black,
-            fontFamily: 'halter',
-            fontSize: 16,
-            package: 'flutter_credit_card',
-          ),
-        );
-
-    return Container(
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8),
-        boxShadow: const <BoxShadow>[
-          BoxShadow(
-            color: Colors.black26,
-            offset: Offset(0, 0),
-            blurRadius: 24,
-          ),
-        ],
-        gradient: backgroundGradientColor,
-      ),
-      margin: const EdgeInsets.all(16),
-      width: widget.width ?? width,
-      height: widget.height ??
-          (orientation == Orientation.portrait ? height / 4 : height / 2),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Expanded(
-            flex: 2,
-            child: Container(
-              margin: const EdgeInsets.only(top: 16),
-              height: 48,
-              color: Colors.black,
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Container(
-              margin: const EdgeInsets.only(top: 16),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: <Widget>[
-                  Expanded(
-                    flex: 9,
-                    child: Container(
-                      height: 48,
-                      color: Colors.white70,
-                    ),
-                  ),
-                  Expanded(
-                    flex: 3,
-                    child: Container(
-                      color: Colors.white,
-                      child: Padding(
-                        padding: const EdgeInsets.all(5),
-                        child: Text(
-                          _cvvController.text.isEmpty
-                              ? isAmex ? 'XXXX' : 'XXX'
-                              : _cvvController.text,
-                          maxLines: 1,
-                          style: widget.textStyle ?? defaultTextStyle,
-                        ),
-                      ),
-                    ),
-                  )
-                ],
-              ),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.bottomRight,
-              child: Padding(
-                padding: const EdgeInsets.only(left: 16, right: 16, bottom: 16),
-                child: getCardTypeIcon(_numberController.text),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  ///
-  /// Builds a front container containing
-  /// Card number, Exp. year and Card holder name
-  ///
-  Container buildFrontContainer(
-    double width,
-    double height,
-    BuildContext context,
-    Orientation orientation,
-  ) {
-    final TextStyle defaultTextStyle =
-        Theme.of(context).textTheme.title.merge(textCardStyle);
-
-    return Container(
-      margin: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8),
-        boxShadow: const <BoxShadow>[
-          BoxShadow(
-            color: Colors.black26,
-            offset: Offset(0, 0),
-            blurRadius: 24,
-          )
-        ],
-        gradient: backgroundGradientColor,
-      ),
-      width: widget.width ?? width,
-      height: widget.height ??
-          (orientation == Orientation.portrait ? height / 4 : height / 2),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Align(
-            alignment: Alignment.topRight,
-            child: Padding(
-              padding: const EdgeInsets.only(left: 16, right: 16, top: 8),
-              child: getCardTypeIcon(_numberController.text),
-            ),
-          ),
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.only(left: 16),
-              child: Text(
-                _numberController.text.isEmpty || widget.cardNumber == null
-                    ? 'XXXX XXXX XXXX XXXX'
-                    : _numberController.text,
-                style: widget.textStyle ?? defaultTextStyle,
-              ),
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: Padding(
-              padding: const EdgeInsets.only(left: 16),
-              child: Text(
-                _expiryController.text.isEmpty || widget.expiryDate == null
-                    ? 'MM/YY'
-                    : _expiryController.text,
-                style: widget.textStyle ?? defaultTextStyle,
-              ),
-            ),
-          ),
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.only(left: 16, right: 16, bottom: 16),
-              child: Text(
-                _nameController.text.isEmpty || widget.cardHolderName == null
-                    ? 'CARD HOLDER'
-                    : _nameController.text,
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-                style: widget.textStyle ?? defaultTextStyle,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// This function determines the Credit Card type based on the cardPatterns
-  /// and returns it.
-  CardType detectCCType(String cardNumber) {
-    //Default card type is other
-    CardType cardType = CardType.otherBrand;
-
-    if (cardNumber.isEmpty) {
-      return cardType;
-    }
-
-    cardNumPatterns.forEach(
-      (CardType type, Set<List<String>> patterns) {
-        for (List<String> patternRange in patterns) {
-          // Remove any spaces
-          String ccPatternStr =
-              cardNumber.replaceAll(RegExp(r'\s+\b|\b\s'), '');
-          final int rangeLen = patternRange[0].length;
-          // Trim the Credit Card number string to match the pattern prefix length
-          if (rangeLen < cardNumber.length) {
-            ccPatternStr = ccPatternStr.substring(0, rangeLen);
-          }
-
-          if (patternRange.length > 1) {
-            // Convert the prefix range into numbers then make sure the
-            // Credit Card num is in the pattern range.
-            // Because Strings don't have '>=' type operators
-            final int ccPrefixAsInt = int.parse(ccPatternStr);
-            final int startPatternPrefixAsInt = int.parse(patternRange[0]);
-            final int endPatternPrefixAsInt = int.parse(patternRange[1]);
-            if (ccPrefixAsInt >= startPatternPrefixAsInt &&
-                ccPrefixAsInt <= endPatternPrefixAsInt) {
-              // Found a match
-              cardType = type;
-              break;
-            }
-          } else {
-            // Just compare the single pattern prefix with the Credit Card prefix
-            if (ccPatternStr == patternRange[0]) {
-              // Found a match
-              cardType = type;
-              break;
-            }
-          }
-        }
-      },
-    );
-
-    return cardType;
-  }
-
-  // This method returns the icon for the visa card type if found
-  // else will return the empty container
-  Widget getCardTypeIcon(String cardNumber) {
-    Widget icon;
-    switch (detectCCType(cardNumber)) {
-      case CardType.visa:
-        icon = Image.asset(
-          'icons/visa.png',
-          height: 48,
-          width: 48,
-          package: 'flutter_credit_card',
-        );
-        isAmex = false;
-        break;
-
-      case CardType.americanExpress:
-        icon = Image.asset(
-          'icons/amex.png',
-          height: 48,
-          width: 48,
-          package: 'flutter_credit_card',
-        );
-        isAmex = true;
-        break;
-
-      case CardType.mastercard:
-        icon = Image.asset(
-          'icons/mastercard.png',
-          height: 48,
-          width: 48,
-          package: 'flutter_credit_card',
-        );
-        isAmex = false;
-        break;
-
-      case CardType.discover:
-        icon = Image.asset(
-          'icons/discover.png',
-          height: 48,
-          width: 48,
-          package: 'flutter_credit_card',
-        );
-        isAmex = false;
-        break;
-
-      default:
-        icon = Container(
-          height: 48,
-          width: 48,
-        );
-        isAmex = false;
-        break;
-    }
-
-    return icon;
-  }
-}
-
-class AnimationCard extends StatelessWidget {
-  const AnimationCard({
-    @required this.child,
-    @required this.animation,
-  });
-
-  final Widget child;
-  final Animation<double> animation;
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedBuilder(
-      animation: animation,
-      builder: (BuildContext context, Widget child) {
-        final Matrix4 transform = Matrix4.identity();
-        transform.setEntry(3, 2, 0.001);
-        transform.rotateY(animation.value);
-        return Transform(
-          transform: transform,
-          alignment: Alignment.center,
-          child: child,
-        );
-      },
-      child: child,
     );
   }
 }
