+++ lib/flutter_credit_card.dart
@@ -1,3 +1,4 @@
 library flutter_credit_card;
 
+export 'package:flutter_credit_card/model/card_field_controller.dart';
 export 'credit_card_widget.dart';
