+++ lib/widget/animation_card.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class AnimationCard extends StatelessWidget {
+  const AnimationCard({
+    @required this.child,
+    @required this.animation,
+  });
+
+  final Widget child;
+  final Animation<double> animation;
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedBuilder(
+      animation: animation,
+      builder: (BuildContext context, Widget child) {
+        final Matrix4 transform = Matrix4.identity();
+        transform.setEntry(3, 2, 0.001);
+        transform.rotateY(animation.value);
+        return Transform(
+          transform: transform,
+          alignment: Alignment.center,
+          child: child,
+        );
+      },
+      child: child,
+    );
+  }
+}
